Int32 is int not long as we are working in both 32 and 64 bits
git-svn-id: http://svn.macosforge.org/repository/ruby/MacRuby/branches/experimental@1004 23306eb0-4c56-4727-a40e-e92c0eb68959

+++ roxor.cpp
@@ -956,7 +956,7 @@ RoxorCompiler::compile_block_create(NODE *node)
     params.push_back(compile_const_pointer(current_block_node));
     params.push_back(current_self);
 
-    params.push_back(ConstantInt::get(Type::Int32Ty, (long)dvars.size()));
+    params.push_back(ConstantInt::get(Type::Int32Ty, (int)dvars.size()));
 
     std::map<ID, Value *>::iterator iter = dvars.begin();
     while (iter != dvars.end()) {
